requirements.txt
pillow~=10.2.0
torch~=2.1.2
torchvision~=0.16.2
opencv-python~=4.9.0.80
Flask~=3.0.0
gunicorn
Flask-Celery
mysql-connector-python~=8.2.0
moviepy~=1.0.3
requests~=2.31.0